+++ sphinxcontrib/sphinx_bazel/domain/domain.py
@@ -9,11 +9,11 @@ Domain overview
  * **package**: contains a ``BUILD`` file. Container for multiple targets
 
   * **target**: any file or .bzl-file. If last, following is available:
-  
+
    * **rule**:  a defined rule inside a .bzl-file
    * **macro**: a defined macro-function inside a .bzl file
-   
-   
+
+
 Domain directives
 =================
 
@@ -41,33 +41,6 @@ from sphinxcontrib.sphinx_bazel.domain.object import BazelObject
 class BazelDomain(Domain):
     """Bazel langage domain"""
 
-    def merge_domaindata(self, docnames, otherdata):
-        for fullname, (fn, objtype) in otherdata['objects'].items():
-            if fn in docnames:
-                self.data['objects'][fullname] = (fn, objtype)
-        for modname, data in otherdata['modules'].items():
-            if data[0] in docnames:
-                self.data['modules'][modname] = data
-
-    def resolve_any_xref(self, env, fromdocname, builder, target, node, contnode):
-        workspace_name = node.get('bazel:workspace')
-        package_name = node.get('py:class')
-        results = []  # type: List[Tuple[str, nodes.Element]]
-    
-        # always search in "refspecific" mode with the :any: role
-        matches = self.find_obj(env, workspace_name, package_name, target, None, 1)
-        for name, obj in matches:
-            if obj[1] == 'module':
-                results.append(('py:mod',
-                                self._make_module_refnode(builder, fromdocname,
-                                                          name, contnode)))
-            else:
-                results.append(('py:' + self.role_for_objtype(obj[1]),
-                                make_refnode(builder, fromdocname, obj[0], name,
-                                             contnode, name)))
-
-            return results
-
     name = 'bazel'
     label = 'Bazel'
     object_types = {
@@ -76,7 +49,8 @@ class BazelDomain(Domain):
         'target': ObjType(l_('target'), 'target', 'ref'),
         'rule': ObjType(l_('rule'), 'rule', 'ref'),
         'macro': ObjType(l_('macro'), 'macro', 'ref'),
-        'impl': ObjType(l_('impl'), 'impl', 'ref')
+        'impl': ObjType(l_('impl'), 'impl', 'ref'),
+        'attribute': ObjType(l_('attribute'), 'attribute', 'ref')
     }
     directives = {
         'workspace': BazelWorkspace,
@@ -86,6 +60,7 @@ class BazelDomain(Domain):
         'macro': BazelObject,
         'implementation': BazelObject,
         'impl': BazelObject,
+        'attribute': BazelObject,
     }
     roles = {}
     initial_data = {
@@ -93,6 +68,6 @@ class BazelDomain(Domain):
         'workspaces': {}
     }
     indices = []
-    
+
     def clear_doc(self, docname):
         pass