+++ docs/source/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 0.1.1
 -----
 
+* Improvement: :ref:`option_name` added to define specific workspace names
 * Improvement: :ref:`option_path` added to support documentation of Bazel objects without a valid workspace
   (`#6 <https://github.com/useblocks/sphinx-bazel/issues/6>`_)
 * Bugfix: Better warnings, if package not found (`#13 <https://github.com/useblocks/sphinx-bazel/issues/13>`_)